+++ sonar-plugin/src/main/java/io/ecocode/javascript/CheckList.java
@@ -34,6 +34,7 @@ public class CheckList {
 
     public static List<Class<? extends JavaScriptCheck>> getAllChecks() {
         return Arrays.asList(
+                AvoidBrightnessOverride.class,
                 AvoidCSSAnimations.class,
                 AvoidHighAccuracyGeolocation.class,
                 LimitDbQueryResult.class,